Add tests for plot.histogram and plot.distributions
